更新.

+++ dan-dome-x/dan_library/src/main/java/com/dan/library/exception/CrashCatchHandler.java
@@ -4,8 +4,10 @@ import android.content.Context;
 import android.os.Environment;
 import android.util.Log;
 
+import com.dan.library.util.ActivityUtil;
 import com.dan.library.util.DateUtil;
 import com.dan.library.util.FileUtil;
+import com.xuexiang.xutil.tip.ToastUtils;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -19,17 +21,41 @@ import java.util.List;
  * Created by Dan on 2019/2/20 11:43
  */
 public class CrashCatchHandler implements Thread.UncaughtExceptionHandler {
+
     private static final String TAG = "CrashCatchHandler";
-    private static CrashCatchHandler INSTANCE;
+
+    /**
+     * 系统默认的UncaughtException处理类
+     */
+    private Thread.UncaughtExceptionHandler mDefaultHandler;
+
     private Context mContext;
-    private static final String PATH = Environment.getExternalStorageDirectory().getPath() + "/log/";
+
     private String thisPath;
+
+    private static CrashCatchHandler INSTANCE;
+
+    private static final String PATH = Environment.getExternalStorageDirectory().getPath() + "/log/";
 
     private CrashCatchHandler() {
 
     }
 
+    public static CrashCatchHandler getInstance() {
+        if (null == INSTANCE) {
+            synchronized (CrashCatchHandler.class) {
+                if (null == INSTANCE) {
+                    INSTANCE = new CrashCatchHandler();
+                }
+            }
+        }
+        return INSTANCE;
+    }
+
     public void init(Context c, String childPath) {
+        // 获取系统默认的UncaughtException处理器
+        mDefaultHandler = Thread.getDefaultUncaughtExceptionHandler();
+        // 设置该CrashHandler为程序的默认处理器
         Thread.setDefaultUncaughtExceptionHandler(INSTANCE);
         //这里的c应该是全局的。
         if (c == null) {
@@ -54,21 +80,10 @@ public class CrashCatchHandler implements Thread.UncaughtExceptionHandler {
         this.init(c, null);
     }
 
-    public static CrashCatchHandler getInstance() {
-        if (null == INSTANCE) {
-            synchronized (CrashCatchHandler.class) {
-                if (null == INSTANCE) {
-                    INSTANCE = new CrashCatchHandler();
-                }
-            }
-        }
-        return INSTANCE;
-    }
-
     /**
      * 把错误日志写到文件中。
      */
-    private void writeToFile(Throwable ex) {
+    private void writeToSDCardFile(Throwable ex) {
         //判断SD卡不存在或无法使用
         if (!Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED)) {
             Log.e(TAG, "SD卡不存在日志无法写入!error:" + ex.getMessage());
@@ -131,12 +146,33 @@ public class CrashCatchHandler implements Thread.UncaughtExceptionHandler {
     }
 
     @Override
-    public void uncaughtException(Thread t, Throwable e) {
+    public void uncaughtException(Thread thread, Throwable ex) {
+        if (!handleException(ex) && mDefaultHandler != null) {
+            // 如果用户没有处理则让系统默认的异常处理器来处理
+            mDefaultHandler.uncaughtException(thread, ex);
+        } else {
+            ToastUtils.toast("很抱歉,程序出现异常,请重新打开!");
+            //退出程序,finish所有的Activity
+            ActivityUtil.getInstance().finishAll();
+            //杀死这个进程避免类似ios直接退出。
+            android.os.Process.killProcess(android.os.Process.myPid());
+            //退出Jvm,释放所占内存资源，0表示正常退出，非0表示异常
+            System.exit(1);
+        }
+    }
+
+    /**
+     * 自定义错误处理,收集错误信息 发送错误报告等操作均在此完成.
+     *
+     * @param ex
+     * @return true:如果处理了该异常信息;否则返回false.
+     */
+    private boolean handleException(Throwable ex) {
+        if (ex == null) {
+            return false;
+        }
         //写入本地日志
-        writeToFile(e);
-        //退出Jvm,释放所占内存资源，0表示正常退出
-        System.exit(0);
-        //杀死这个进程避免类似ios直接退出。
-        android.os.Process.killProcess(android.os.Process.myPid());
+        writeToSDCardFile(ex);
+        return true;
     }
 }
